refactor(ui): extract QuickModeForm/FullModeForm and integrate useMunicipalitiesLoader/useRentDeclineHint into InvestmentForm
